refactor: split up hooks
