+++ packages/connect/src/clients/createAppClient.test.ts
@@ -3,7 +3,6 @@ import { viem } from "./ethereum/viem";
 
 describe("createAppClient", () => {
   const config = {
-    relayURI: "https://connect.farcaster.xyz",
     ethereum: viem(),
   };
